rag_vault/pyproject.toml
[project]
name = "rag-vault"
version = "0.1.0"
description = "RAG search for the vault directory"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=0.2.0",
    "python-dotenv>=1.0.0",
    "rank-bm25>=0.2.2",
    "pydantic>=2.0.0",
    "streamlit>=1.35.0",
    "langgraph>=0.2.0",
    "pandas>=2.0.0",
]

[tool.uv]
managed = true
